pyproject.toml
[tool.poetry]
name = "wtf"
version = "0.1.0"
description = ""
authors = ["Psy-Crow <72886573+psy-crow-wow@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
